fix: suppress warning for http gateway about health check
